Fix types, run typechecker in CI (#1393)
* python: add trivial type casts to make typechecker happy


* worker.cancel: don't crash on missing child


* webhook_caller_filtered: expects as filter set()


* cog.server.http: don't crash when cpu_count couldn't be determined


* ast_openapi_schema: deal with bytes() defaults


* cog.server.http: change PredictionRequest to make pyright happy


* cog.predictor: more possible types


* Fix: UnionType is new in 3.10


* Fix: | syntax is only available in 3.10


* cog.types: fix untyped definition


* Add tool.pyright config to pyproject.toml


* python: error on missing type arguments, add them everywhere


* ast_openapi_schema.extract_info: make type checker actually happy


* Add typecheck python github action


* cog.server.http: quote typenames


* python: process some review comments


* async runner/http: fix types


* python: mypy -> pyright everywhere


* pyproject: unneccesary -> unnecessary


* resolve threading|asyncio.Event difference the other way


* python: make tests pass for PredictionResponse


* Fix test in python 3.7


---------

+++ python/cog/server/worker.py
@@ -88,7 +88,7 @@ class Worker:
             self._child.join()
 
     def cancel(self) -> None:
-        if self._allow_cancel and self._child.is_alive():
+        if self._allow_cancel and self._child.is_alive() and self._child.pid is not None:
             os.kill(self._child.pid, signal.SIGUSR1)
             self._allow_cancel = False
